requirements.txt
python-dotenv
openai # Used for DeepSeek API interaction structure
python-docx
chromadb
tiktoken # Often used by embedding models/LLMs for token counting
matplotlib
plotly # Optional, for potentially richer visualizations
pydantic
# reportlab # Optional: for PDF generation from scratch
# pandoc # Optional: alternative for DOCX -> PDF conversion (requires separate installation)
# numpy # Often a dependency for embedding/vector libs
sentence-transformers
torch
transformers